Delete `execution.execution` module.
Use `await WorkflowExecution(...)`
