+++ virtool_workflow/__init__.py
@@ -2,7 +2,7 @@
 from virtool_workflow.fixtures import FixtureScope, fixture
 from virtool_workflow.execution import hooks
 from virtool_workflow.execution.hooks.hooks import hook, Hook
-from virtool_workflow.execution.workflow_executor import WorkflowExecution, State, WorkflowError
+from virtool_workflow.execution.workflow_execution import WorkflowExecution, State, WorkflowError
 from virtool_workflow.workflow import Workflow
 from virtool_workflow.decorator_api import step, cleanup, startup
