+++ virtool_workflow/execution/hooks/workflow_hooks.py
@@ -1,4 +1,5 @@
 from virtool_workflow.execution.hooks import FixtureHook
+from virtool_workflow.execution.workflow_execution import State
 
 on_result = FixtureHook("on_result", [], None)
 """
@@ -50,3 +51,36 @@ Triggered on a change of state during workflow execution.
     async def do_something_on_state_change(old_state, new_state):
         ...
 """
+
+on_workflow_failure = FixtureHook("on_workflow_finish", [Exception], None)
+"""
+Triggered when a workflow fails to complete.
+
+.. code-block:: python
+
+    @on_on_workflow_failure
+    async def perform_on_failure(cause: Exception, execution: WorkflowExecution):
+        ...
+"""
+
+on_workflow_finish = FixtureHook("on_workflow_finish", [], None)
+"""
+Triggered when a workflow finishes, regardless of it's success.
+
+.. code-block:: python
+
+    @on_workflow_finish
+    async def perform_on_success(workflow: Workflow):
+        ...
+"""
+
+
+@on_workflow_failure
+async def _trigger_finish_from_failure(_, scope):
+    await on_workflow_finish.trigger(scope)
+
+
+@on_result
+async def _trigger_finish_from_success(scope):
+    await on_workflow_finish.trigger(scope)
+
